move decode tests to decode_test.go
